backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic>=2.10,<3
pydantic-settings>=2.6,<3
email-validator>=2.2,<3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
psycopg2-binary==2.9.9

# AI/ML
openai==1.10.0
scikit-learn==1.4.2
joblib==1.4.2

# Security & API Integration
requests==2.31.0
httpx==0.26.0
aiohttp==3.9.1

# CORS
fastapi-cors==0.0.6
